Fixed the variable_length_mnist test method, which got broken because of missing new arguments.
Found a weird bug in multi_dimensional_lstm_parameters:
the class MultiDirectionalMultiDimensionalLSTMParametersFullyParallel
has a field "next_input_column_index". This field is set to zero in the
constructor, and incremented by the method "prepare_computation_next_column_functions".
However, it is never being reset! One would expect this to manifest in a bug, whereby
the index, which is used to determine the column in the input tensor, gets
out of bounds. When NOT using data parallel, this is indeed what happens.
But somehow, use of (custom) data parallel has masked this bug, since for some
reason the field gets magically reset in the right way after each batch, when using
data parallel. While this is really weird, it is of course not a functionality one
should rely upon. Hence this is fixed now with an explicit reset to zero in
the multi_dimensional_lstm.py class. It is still a mystery why by accident it was
working in the intended way when using data parallel: does data parallel somehow
by itself reset certain primitive fields of Modules? No idea. It is really odd.
Tested it also with another such integer variable. Same behavior: with using
data parallel it gets reset after every epoch, without it, it keeps increasing
as one would actually expect to happen.

	modified:   custom_data_parallel/scatter_gather.py
	modified:   modules/mdlstm_examples_packing.py
	modified:   modules/multi_dimensional_lstm.py
	modified:   modules/multi_dimensional_lstm_layer_pair_stacking.py
	modified:   modules/multi_dimensional_lstm_parameters.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   util/tensor_list_chunking.py