Fixed again the older MDLSTM computation, with only one-directional computation and no examples packing. As it turns out, it seems that looking at top, this combination does not leak memory; or if it does it is barely noticable. This in contrast to four-directional with examples packing. Furthermore, using one GPU (without data parallel) or two gpus (with custom data parallel) does not make a difference, as long as only one-directional computation and no examples packing is used, no memory appears to be leaking - based on observing top.
The memory leaking also happens even when only one GPU is used without
data parallel, if four-directional computation and packing is used.

This suggests the memory leak is not in the custom data parallel code,
but rather in the four-directional computation and/or examples packing
code. There are still some other options as well, as there are other
methods that are only used with four-directional computation or with
packing. In either way, this narrows it down a bit, and also seems
to confirm it is really a python memory leak problem, and not in
c code such as the ctc-decode module.

To narrow it down even further, four-directional computation but without
examples packing might be good to get working, in order to see
if this still works without memory leak.

	modified:   modules/mdlstm_layer_block_strided_convolution_layer_pair.py
	modified:   modules/multi_dimensional_lstm.py
	modified:   modules/multi_dimensional_lstm_layer_pair_stacking.py
	modified:   modules/multi_dimensional_lstm_parameters.py
	modified:   modules/multi_dimensional_rnn.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py