Added a class tests/test_multi_dimensional_lstm_parameters.py aimed at isolating the cause of the appearant memory leak. The leak was then isolated to "mdlstm_parameters.prepare_computation_next_column_functions" and from there to the function "compute_convolution_result_and_apply_mask"  in the class "ParallelMultipleStateWeightingsComputation" . Essentially the memory leak just happens when the Conv2d forward function is called within this class, or so it appears.
Whereas it is not clear whether this happens because of too many
groups (28 * 2 = 56), it appears this bug has only appeared with
the fully parallel MDLSTM computation, which combines everything
into a single convolutional computation, as much as possible.
Going back to a slightly less parallelized version of the code,
might hence be a workaround that fixes the memory leak, though
not a very satisfactory one of course. What remains to be seen
perhaps is whether the bug can be reproduced with even simpler code,
and based on that a bug could be posted on the pytorch forum, to
get more help in resolving the problem.

Added comments to indicate where the memory leak seems to reside
(based on experimentation commenting out parts of the code, and
checking if the memory increases or not using "top" in linux).

	modified:   modules/mdlstm_examples_packing.py
	modified:   modules/multi_dimensional_lstm_parameters.py
	modified:   modules/parallel_multiple_state_weightings_computation.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	new file:   tests/test_multi_dimensional_lstm_parameters.py