Further extended the code aimed at finding the memory leak plaguing the current code, and which appears it seems when examples packing is used. A mdlstm network without block-strided convolution layers was added to enable excluding the block-strided convolution as a cause of the memory leak. Also the test code for mdlst_examples_packing by itself was extended, so far strangely enough, the memory leak cannot be reproduced when examples packing is used by itself.
The code was furthermore fixed so that examples-packing can be used without
four-directional computation. This still gives a memory leak it seems, which
suggest the problem is located in the examples packing, not in the four-directional
computation.

As it is, it remains difficult to decide even with certainty whether or not the
memory leak is occuring or not, using top. While not solving the poblem, these
changes do help narrowing down the likely causes somewhat.

	modified:   modules/mdlstm_examples_packing.py
	modified:   modules/multi_dimensional_lstm_layer_pair_stacking.py
	modified:   modules/network_to_softmax_network.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py
	modified:   tests/test_multi_dimensional_lstm_parameters.py