Fixed the combination of multi-directional computation and no examples packing, which was not working previously. This required changes in network_to_softmax_network, which was not resizing the activations in the right way to form the 2D input to the convolution, in case multiple directions were used without packing. It furthermore required changes in multi_dimensional_lstm, preparing a four-directional input, as was done previously only when examples packing was being used.
This fix now should allow to run the network with four-directional computation and with
or without examples packing, allowing to compare the runtimes of both, and thereby
determine the speed benefit of using examples packing. Note that the speed benefits
are mainly caused by saving out memory with examples packing, allowing a larger
batch size to be used with the same memory and thereby increasing speed. For the
comparison both models (with or without examples packing) should be ran with the maximum
batch size that still works for that setting.

	modified:   modules/block_strided_convolution.py
	modified:   modules/mdlstm_examples_packing.py
	modified:   modules/multi_dimensional_lstm.py
	modified:   modules/multi_dimensional_lstm_parameters.py
	modified:   modules/network_to_softmax_network.py