Added a required specification of whether or not to use example packing through configuration arguments. This is useful for the new experiments where the memory usage and run-time is compared with and without example packing. Also added a consistency check of the collate function used by data loader, which may be inconsistent when the data loader was created for use with example packing, but then re-used in a setting without example packing.
The tricky part is that the data loader must do last-minute-padding, using the custom collate function,
if no packing is being used. This cannot be done inside the network, for there there is no way to
to coordinate the maximum widht and height within the batch, when the batch is split between multiple
gpus.

Perhaps another way would be to update the custom collate function of the data loader.

	modified:   modules/opts.py
	modified:   modules/train_multi_dimensional_rnn_ctc.py